+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.1.3
 * Support for V9
+* Distraction type now only shows on distraction decorators
 
 ### 1.1.2
 * Fixed chat not scrolling to bottom when zone generator adds the clear button